refactor: extract affected jobs columns to dedicated hook
Moves column definitions from DeprecationAffectedJobs component into a
useDeprecationAffectedJobsColumns hook to follow codebase conventions
and resolve SonarQube nested component warning.
