+++ frontend/awx/administration/deprecations/DeprecationAffectedJobs.tsx
@@ -1,21 +1,16 @@
 import {
-  DateTimeCell,
-  ITableColumn,
   IToolbarFilter,
   PageTable,
-  TextCell,
   ToolbarFilterType,
-  useGetPageUrl,
   useInMemoryView,
 } from '@ansible/ansible-ui-framework';
-import { StatusCell } from '@ansible/common-ui/Status';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import useSWR from 'swr';
 import { requestGet } from '@ansible/common-ui/crud/Data';
 import { awxAPI } from '../../common/api/awx-utils';
-import { AwxRoute } from '../../main/AwxRoutes';
+import { useDeprecationAffectedJobsColumns } from './hooks/useDeprecationAffectedJobsColumns';
 import { useDeprecationData } from './hooks/useDeprecationData';
 
 interface AffectedJob {
@@ -33,7 +28,6 @@ interface AffectedJob {
 
 export function DeprecationAffectedJobs() {
   const { t } = useTranslation();
-  const getPageUrl = useGetPageUrl();
   const { deprecationType } = useParams<{ deprecationType: string }>();
   const decodedType = decodeURIComponent(deprecationType ?? '');
 
@@ -57,56 +51,7 @@ export function DeprecationAffectedJobs() {
     }));
   }, [jobsResponse, deprecation]);
 
-  const columns = useMemo<ITableColumn<AffectedJob>[]>(
-    () => [
-      {
-        header: t('ID'),
-        cell: (job) => <TextCell text={String(job.id)} />,
-        sort: 'id',
-        minWidth: 0,
-        card: 'hidden',
-        list: 'hidden',
-      },
-      {
-        header: t('Name'),
-        cell: (job) => (
-          <TextCell
-            text={job.summary_fields.job_template?.name ?? job.name}
-            to={getPageUrl(AwxRoute.JobDetails, { params: { id: job.id, job_type: job.type } })}
-          />
-        ),
-        sort: 'name',
-        card: 'name',
-        list: 'name',
-      },
-      {
-        header: t('Status'),
-        cell: (job) => <StatusCell status={job.status} />,
-        sort: 'status',
-        card: 'subtitle',
-        list: 'subtitle',
-      },
-      {
-        header: t('Occurrences'),
-        cell: (job) => <TextCell text={String(job.occurrences)} />,
-        sort: 'occurrences',
-        minWidth: 0,
-      },
-      {
-        header: t('Started'),
-        cell: (job) => <DateTimeCell value={job.started || undefined} />,
-        sort: 'started',
-      },
-      {
-        header: t('Finished'),
-        cell: (job) => <DateTimeCell value={job.finished || undefined} />,
-        sort: 'finished',
-        defaultSort: true,
-        defaultSortDirection: 'desc',
-      },
-    ],
-    [t, getPageUrl]
-  );
+  const columns = useDeprecationAffectedJobsColumns();
 
   const toolbarFilters = useMemo<IToolbarFilter[]>(
     () => [
